Separate out output dir builder as well
