+++ src/translator.py
@@ -1,9 +1,10 @@
 import os
 import polib
 
-from Helpers import create_pot_file, create_pot_file_from_keyed, create_languagedata_xml_file, create_logger, merge_compendium
+from Helpers import create_pot_file_from_keyed, create_logger
 from Parser import parse_arguments
 from source_dir_builder import SourceDirBuilder
+from output_dir_builder import OutputDirBuilder
 
 class Translator:
     def __init__(self):
@@ -22,7 +23,7 @@ class Translator:
         if self.args.source_dir:
             SourceDirBuilder(self.args, self.compendium, self.logger).build_source_dir()
         if self.args.output_dir:
-            self.build_output_dir()
+            OutputDirBuilder(self.args, self.logger).build_output_dir()
 
     def build_compendium(self):
         if self.compendium is not None:
@@ -45,44 +46,3 @@ class Translator:
                     compendium += create_pot_file_from_keyed(full_filename, self.args.source_dir, self.args.compendium, True)
 
         return compendium
-
-    def build_output_dir(self):
-        self.logger.info('Beginning to generate DefInjected files')
-        fuzzy = 0
-        total = 0
-        translated = 0
-        untranslated = 0
-
-        for root, dirs, files in os.walk(self.args.po_dir):
-            for file in files:
-                if file.endswith('.po'):
-                    full_filename = os.path.join(root, file)
-                    self.logger.info("Processing " + full_filename)
-                    xml_filename = full_filename.split(self.args.po_dir, 1)[1]
-                    xml_filename = xml_filename.strip('.po')
-                    xml_filename = os.path.join(self.args.output_dir, xml_filename)
-                    directory = os.path.dirname(xml_filename)
-
-                    po = polib.pofile(full_filename)
-                    translated_po_entries = len(po.translated_entries())
-                    fuzzy_po_entries = len(po.fuzzy_entries())
-                    untranslated_po_entries = len(po.untranslated_entries())
-
-                    translated = translated + translated_po_entries
-                    fuzzy = fuzzy + fuzzy_po_entries
-                    untranslated = untranslated + untranslated_po_entries
-
-                    # Do we have translated entries?
-                    if translated_po_entries > 0:
-                        if not (os.path.exists(directory)):
-                            self.logger.info("Creating directory " + directory)
-                            os.makedirs(directory)
-                        xml_content = create_languagedata_xml_file(full_filename)
-                        target = open(xml_filename, "w", encoding="utf8")
-                        target.write(xml_content)
-                        target.close()
-                    total_po_entries = len([e for e in po if  not e.obsolete])
-                    total = total + total_po_entries
-
-        print("Statistics (untranslated/fuzzy/translated/total): %d/%d/%d/%d" % (untranslated, fuzzy, translated, total))
-
